delete files, dates, post sound

+++ delete.php
@@ -1,11 +1,18 @@
 <?php
-// http:delete?file=abc.def
-// deletes file files/abc.def
+// http:delete?file=directory/abc.def
+// deletes file files/directory/abc.def
+// and directory
 // Written by Wm Leler
 
+if ($_GET['file'] === NULL) { return; }
 $file = htmlspecialchars($_GET['file']);
 $filepath = 'files/'.$file;
 if (unlink($filepath)) {
-	echo $filepath.' deleted';
+	echo $filepath.' file deleted<br>';
+}
+$subs = explode('/', $file);
+$filepath = 'files/'.$subs[0];
+if (count($subs) === 2 && rmdir($filepath)) {
+	echo $filepath.' directory removed<br>';
 }
 ?>
